update: Added swap which makes me cry.
- Added some functions.
- Refactored.
