+++ example/example.ino
@@ -1,4 +1,4 @@
-#include <sorter.h>
+#include <utils.h>
 
 void setup() {
     Serial.begin(9600);
@@ -12,11 +12,11 @@ void loop() {
             a[i][0] = random(100);
             a[i][1] = i;
         }
-  
-        //  sorter::sort_max(a); // Set two first indexes in array to max values
-        sorter::sort_min(a); // Set two first indexes in array to min values
-        //  sorter::swap(); // Swaps two numbers 
-        
+
+        utils::sort_max(a);  // Set two first indexes in array to max values
+        utils::sort_min(a);  // Set two first indexes in array to min values
+        utils::swap();       // Swaps two numbers
+
         for (int i = 0; i < 12; i++) {
             Serial.print(a[i][1]);
             Serial.print(' ');
