--- sorter.h
@@ -1,11 +1,0 @@
-#ifndef SORTER_h
-#define SORTER_H
-
-namespace sorter {
-    void sort_max(int s[12][2]);
-    void sort_min(int s[12][2]);
-    void swap(int *a, int *b);
-}
-
-#endif // SORTER_H
-
